requirements.txt
langgraph
langchain
langchain_openai
pydantic
psycopg2-binary
redis
pytest
python-dotenv
